requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
boto3==1.40.32
botocore==1.40.32
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.1
dataclasses-json==0.6.7
debugpy==1.8.17
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.116.2
fastapi-cli==0.0.12
fastapi-cloud-cli==0.1.5
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.28
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.3
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-telegram-bot==22.4
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
ruff==0.13.0
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.0
sentry-sdk==2.38.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
tqdm==4.67.1
transformers==4.56.1
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uv==0.8.17
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zstandard==0.25.0
